feat: add `deleteOriginalPages` option

+++ packages/gatsby-plugin-intl/src/gatsby-node.js
@@ -9,11 +9,17 @@ export const onCreatePage = async (
   { page, actions: { createPage, deletePage, createRedirect } },
   pluginOptions
 ) => {
-  const isEnvDevelopment = process.env.NODE_ENV === 'development';
   const { siteUrl, supportedLanguages, defaultLanguage, notFoundPage, excludedPages } = {
     ...DEFAULT_OPTIONS,
     ...pluginOptions,
   };
+
+  const isEnvDevelopment = process.env.NODE_ENV === 'development';
+  const deleteOriginalPages =
+    pluginOptions.deleteOriginalPages !== undefined
+      ? pluginOptions.deleteOriginalPages
+      : !isEnvDevelopment;
+
   const originalPath = page.path;
   const is404 = originalPath.includes(notFoundPage);
 
@@ -24,7 +30,9 @@ export const onCreatePage = async (
 
   // Delete the original page (since we are gonna create localized versions of it) and add a
   // redirect header
-  await deletePage(page);
+  if (deleteOriginalPages) {
+    await deletePage(page);
+  }
 
   await Promise.all(
     supportedLanguages.map(async lang => {
